rename : 폴더 명 fixzone -> fix 로 변경
