+++ src/components/fix/Clearbutton.tsx
