+++ src/hooks/api/fixzone/useClubFixInfo.ts
@@ -1,7 +1,7 @@
 import { useQuery } from '@tanstack/react-query';
 import { AxiosError, type AxiosResponse } from 'axios';
 import { getClubFixInfo } from '@/apis';
-import { FixClubDetailType } from '@/types/fixzone';
+import { FixClubDetailType } from '@/types/fix';
 type Props = {
   token: string;
   id: number;
