+++ src/hooks/api/fixzone/useMyFix.ts
@@ -1,7 +1,7 @@
 import { useQuery } from '@tanstack/react-query';
 import { AxiosError, type AxiosResponse } from 'axios';
 import { getClubAllFix } from '@/apis';
-import { Fix } from '@/types/fixzone';
+import { Fix } from '@/types/fix';
 
 export function useMyFix(token: string) {
   return useQuery<unknown, AxiosError, AxiosResponse<Fix[], unknown>, [string]>(
