+++ src/middleware.ts
@@ -45,8 +45,8 @@ export const config = {
     '/my-club',
     '/club',
     '/member',
-    '/fixzone',
-    '/fixzone/:path*',
+    '/fix',
+    '/fix/:path*',
     '/club/:path*',
     '/report',
     '/banner',
